+++ addon/services/groups-service.js
@@ -9,16 +9,20 @@ export default Ember.Service.extend(serviceMixin,{
   search(form){
 
     let qs = this.encodeForm(form);
-
-    let portalBaseUrl = 'https://www.arcgis.com';
-    if(this.get('hostAppConfig.APP.portalBaseUrl')){
-      portalBaseUrl = this.get('hostAppConfig.APP.portalBaseUrl');
-    }
-    portalBaseUrl = portalBaseUrl + '/sharing/rest';
+    let portalRestUrl = this.getPortalRestUrl();
     //q=title%3A${window.encodeURIComponent(q)}+AND+access%3Apublic+AND+isopendata%3Atrue&sortField=title&num=100&f=json
-    let url = `${portalBaseUrl}/community/groups?${qs}`;
+    let url = `${portalRestUrl}/community/groups?${qs}`;
 
     return this.request(url);
-  }
+  },
 
+  /**
+   * Get the item json
+   */
+  getById(groupId){
+    let qs = this.encodeForm({});
+    let portalRestUrl = this.getPortalRestUrl();
+    let url = `${portalRestUrl}/content/groups/${groupId}?${qs}`;
+    return this.request(url);
+  }
 });