+++ CHANGELOG.md
@@ -3,8 +3,12 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+### Added
+- get items by group id
 ### Changes
 - correctly pull in parent app configuration
+- refactor logic for getting portal base URL into mixin
+
 
 ## [0.2.4]
 ### Added